Separate ImportedImage model
